+++ Tests/SwiftInfoTests/MockSourceKit.swift
@@ -1,0 +1,6 @@
+import Foundation
+import SwiftInfoCore
+
+class MockSourceKit: SourceKit {
+
+}